slice: Use int64_t for prefetch block math
Block numbers are int64_t (Range, Data::m_blocknum); widen the
prefetch high-water mark and loop accordingly and narrow only at the
BgBlockFetch::schedule() boundary. Addresses Copilot review.

+++ plugins/slice/Data.h
@@ -94,7 +94,7 @@ struct Data {
 
   bool m_prefetchable{false};
 
-  int m_prefetch_hwm{-1}; // highest block this request has scheduled for prefetch
+  int64_t m_prefetch_hwm{-1}; // highest block this request has scheduled for prefetch
 
   HdrMgr m_req_hdrmgr;  // manager for server request
   HdrMgr m_resp_hdrmgr; // manager for client response
